requirements.txt
pulumi>=3.0.0
pulumi-gcp>=4.0.0
PyYAML>=5.4.1
